requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
distlib==0.3.6
distro==1.8.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
docker==6.0.1
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.26.0
Faker==17.6.0
filelock==3.8.1
idna==3.4
inflection==0.5.1
jsonschema==3.2.0
packaging==21.3
paramiko==2.12.0
pathlib==1.0.1
pipenv==2022.11.30
platformdirs==2.5.4
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.0
virtualenv-clone==0.5.7
websocket-client==0.59.0
